nrf: Enable ubinascii module
These modules would be very handy to have in order to implement a
more robust replacement for ampy.  It costs XXXX bytes of text and
no data or bss.

'ubinascii' has base64 encoding, which can be used to create a fully
8-bit-clean transport above the slightly cooked serial connection of
these boards.

I see this as only being useful on ports which don't have USB block
device support, so atmel-samd doesn't need it.  These modules are
already on esp8266 so I think nrf52832 seems like the only one that
needs it.